Move bootstrap token utilities
extract into own package to resolve cyclic dependency
